Add initial mc xp tipping command handling
