+++ src/interactions/cs/tipXpAttendance.ts
@@ -1,0 +1,218 @@
+import {
+  ChatInputCommandInteraction,
+  EmbedBuilder,
+  GuildMember,
+  MessageContextMenuCommandInteraction,
+  UserContextMenuCommandInteraction,
+  VoiceBasedChannel
+} from 'discord.js';
+import { getAddress } from 'viem';
+
+import {
+  checkUserNeedsCooldown,
+  dropExp,
+  getCharacterAccountsByPlayerAddresses,
+  getPlayerAddressesByDiscordHandles,
+  updateLatestXpTip
+} from '@/lib';
+import { ClientWithCommands } from '@/types';
+import { EXPLORER_URL, RAIDGUILD_GAME_ADDRESS } from '@/utils/constants';
+import { discordLogger } from '@/utils/logger';
+
+export const tipXpAttendanceInteraction = async (
+  client: ClientWithCommands,
+  interaction:
+    | ChatInputCommandInteraction
+    | MessageContextMenuCommandInteraction
+    | UserContextMenuCommandInteraction
+) => {
+  const TIP_AMOUNT = '20';
+  const TABLE_NAME = 'latestAttendanceXpTips';
+  const MINIMUM_ATTENDEES = 6;
+
+  if (!EXPLORER_URL) {
+    discordLogger('Missing EXPLORER_URL env', client);
+    return;
+  }
+
+  const senderId = interaction.user.id;
+  const { endTime, lastSenderDiscordId, needsCooldown } =
+    await checkUserNeedsCooldown(client, TABLE_NAME);
+
+  if (needsCooldown) {
+    const embed = new EmbedBuilder()
+      .setTitle('Attendance XP Tipping Cooldown')
+      .setDescription(
+        `All members must wait ${
+          endTime
+            ? `until ${endTime} to tip attendance again.`
+            : '24 hours between attendance tipping.'
+        } `
+      )
+      .setColor('#ff3864')
+      .setTimestamp();
+
+    await interaction.followUp({
+      embeds: [embed]
+    });
+    return;
+  }
+
+  const channelId = interaction.channel?.id;
+
+  const voiceChannel = interaction.guild?.channels.cache.get(
+    channelId ?? ''
+  ) as VoiceBasedChannel;
+  const { members } = voiceChannel;
+  const discordMembers = members.map(m => m);
+
+  if (discordMembers.length < MINIMUM_ATTENDEES) {
+    const embed = new EmbedBuilder()
+      .setTitle('Not Enough Attendees')
+      .setDescription(
+        `There must be at least ${MINIMUM_ATTENDEES} attendees in the voice channel to tip attendance.`
+      )
+      .setColor('#ff3864')
+      .setTimestamp();
+
+    await interaction.followUp({
+      embeds: [embed]
+    });
+    return;
+  }
+
+  const [senderTagToEthAddressMap] = await getPlayerAddressesByDiscordHandles(
+    client,
+    interaction,
+    [interaction.member as GuildMember]
+  );
+
+  if (
+    !senderTagToEthAddressMap ||
+    !senderTagToEthAddressMap[interaction.user.tag]
+  ) {
+    const embed = new EmbedBuilder()
+      .setTitle('Not a Member')
+      .setDescription(
+        `You are not a member of RaidGuild! If you think this is an error, ensure that your Discord handle and ETH address are registered correctly in DungeonMaster.`
+      )
+      .setColor('#ff3864')
+      .setTimestamp();
+
+    await interaction.followUp({
+      embeds: [embed]
+    });
+    return;
+  }
+
+  const [discordTagToEthAddressMap, discordTagsWithoutEthAddress] =
+    await getPlayerAddressesByDiscordHandles(
+      client,
+      interaction,
+      discordMembers as GuildMember[]
+    );
+
+  if (!discordTagToEthAddressMap) return;
+  const playerAddresses = Object.values(discordTagToEthAddressMap);
+  if (!playerAddresses) return;
+
+  const [discordTagToCharacterAccountMap, discordTagsWithoutCharacterAccounts] =
+    await getCharacterAccountsByPlayerAddresses(
+      client,
+      interaction,
+      discordTagToEthAddressMap
+    );
+  if (!discordTagToCharacterAccountMap) return;
+  const accountAddresses = Object.values(discordTagToCharacterAccountMap);
+  if (!accountAddresses) return;
+
+  if (accountAddresses.length === 0) {
+    const embed = new EmbedBuilder()
+      .setTitle('No Characters Found')
+      .setDescription(
+        `No characters were found for the following users: ${discordMembers.map(
+          m => `<@${m?.id}>`
+        )}.\n---\nIf you think this is an error, ensure that your Discord handle and ETH address are registered correctly in DungeonMaster.`
+      )
+      .setColor('#ff3864')
+      .setTimestamp();
+
+    await interaction.followUp({
+      embeds: [embed]
+    });
+    return;
+  }
+
+  const tx = await dropExp(client, interaction, accountAddresses, TIP_AMOUNT);
+  if (!tx) return;
+
+  const txHash = tx.hash;
+
+  let embed = new EmbedBuilder()
+    .setTitle('Attendance XP Tipping Tx Pending...')
+    .setURL(`${EXPLORER_URL}/tx/${txHash}`)
+    .setDescription(
+      `Transaction is pending. View your transaction here:\n${EXPLORER_URL}/tx/${txHash}`
+    )
+    .setColor('#ff3864')
+    .setTimestamp();
+
+  await interaction.followUp({
+    embeds: [embed]
+  });
+
+  const txReceipt = await tx.wait();
+
+  if (!txReceipt.status) {
+    embed = new EmbedBuilder()
+      .setTitle('Attendance XP Tipping Tx Failed!')
+      .setURL(`${EXPLORER_URL}/tx/${txHash}`)
+      .setDescription(
+        `Transaction failed. View your transaction here:\n${EXPLORER_URL}/tx/${txHash}`
+      )
+      .setColor('#ff3864')
+      .setTimestamp();
+
+    await interaction.editReply({
+      embeds: [embed]
+    });
+    return;
+  }
+
+  const discordMembersSuccessfullyTipped = discordMembers.filter(
+    m =>
+      !discordTagsWithoutCharacterAccounts?.includes(m?.user.tag as string) &&
+      !discordTagsWithoutEthAddress?.includes(m?.user.tag as string)
+  );
+  const discordIdsSuccessfullyTipped = discordMembersSuccessfullyTipped.map(
+    m => m?.user.id
+  );
+
+  embed = new EmbedBuilder()
+    .setTitle('Attendance XP Tipping Succeeded!')
+    .setURL(`${EXPLORER_URL}/tx/${txHash}`)
+    .setDescription(
+      `**<@${senderId}>** tipped ${TIP_AMOUNT} XP to all characters in this voice channe:\n${discordIdsSuccessfullyTipped.map(
+        id => `<@${id}>`
+      )}.\n---\nIf you did not receive a tip, you are either not a member of RaidGuild, not in DungeonMaster, or not in CharacterSheets.`
+    )
+    .setColor('#ff3864')
+    .setTimestamp();
+
+  const gameAddress = getAddress(RAIDGUILD_GAME_ADDRESS);
+
+  const data = {
+    lastSenderDiscordId,
+    newSenderDiscordId: senderId,
+    senderDiscordTag: interaction.user.tag,
+    gameAddress,
+    chainId: '5',
+    txHash
+  };
+
+  await updateLatestXpTip(client, TABLE_NAME, data);
+
+  await interaction.editReply({
+    embeds: [embed]
+  });
+};
