+++ src/interactions/index.ts
@@ -1,5 +1,6 @@
 export {
   tipXpAttendanceInteraction,
-  tipXpInteraction
+  tipXpInteraction,
+  tipXpMcInteraction
 } from '@/interactions/cs';
 export { queryInteraction } from '@/interactions/dm';
